Added refund with credit scaling according to number of days
>= 6 days, 100%
5 days, 80%
3-4 days 50%
1-2days 0%

+++ src/main/java/com/example/cs203g1t3/service/UserService.java
@@ -44,7 +44,7 @@ public interface UserService {
 
     ResponseEntity<?> registerAccount(SignupRequest signUpRequest, Role role);
 
-    void addCreditScore(Long userId,int creditScoreToAdd);
+    void addCreditScore(Long userId,double creditScoreToAdd);
 
     void deductCredit(Long userId, double creditDeducted);
